Factor out AddPhoneNumber to a separate component
